tir: Rearrange terms from `scopes` into `blocks` and `commands`
