+++ compiler/hash-tir/src/tir/terms/control.rs
@@ -4,26 +4,15 @@ use core::fmt;
 use std::fmt::Debug;
 
 use hash_ast::ast::MatchOrigin;
-use hash_storage::store::{statics::StoreId, SequenceStoreKey, TrivialSequenceStoreKey};
+use hash_storage::store::{statics::StoreId, TrivialSequenceStoreKey};
 use textwrap::indent;
 
 use crate::{
     scopes::StackId,
     stores::tir_stores,
-    tir::{PatId, PatListId, Term, TermId},
+    tir::{PatId, PatListId, TermId},
     tir_node_sequence_store_direct,
 };
-
-/// A loop term.
-///
-/// Contains an inner term which should produce side-effects.
-///
-/// The type of a loop is `void`, unless it can be proven to never terminate (in
-/// which case it is `never`).
-#[derive(Debug, Clone, Copy)]
-pub struct LoopTerm {
-    pub inner: TermId,
-}
 
 /// A match term.
 ///
@@ -53,25 +42,6 @@ pub struct MatchCase {
 
 tir_node_sequence_store_direct!(MatchCase);
 
-/// A return term.
-///
-/// This is typed as `never`.
-#[derive(Debug, Clone, Copy)]
-pub struct ReturnTerm {
-    // The expression to return. If there is no such expression in the code, then this will be the
-    // empty tuple term.
-    pub expression: TermId,
-}
-
-/// A term relating to loop control flow.
-///
-/// Either break or continue.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LoopControlTerm {
-    Break,
-    Continue,
-}
-
 /// A conditional pattern, containing a pattern and an condition.
 ///
 /// This is `A if c` in its most general form, where `A` is a pattern and `c` is
@@ -97,12 +67,6 @@ pub struct IfPat {
 pub struct OrPat {
     /// The sequence of alternative patterns.
     pub alternatives: PatListId,
-}
-
-impl fmt::Display for LoopTerm {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "loop {}", self.inner)
-    }
 }
 
 impl fmt::Display for MatchTerm {
@@ -135,26 +99,6 @@ impl fmt::Display for MatchCase {
         let lines = indent(&case, "  ");
         write!(f, "{lines}")?;
         Ok(())
-    }
-}
-
-impl fmt::Display for ReturnTerm {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if matches!(*self.expression.value(), Term::Tuple(tuple_term) if tuple_term.data.value().is_empty())
-        {
-            write!(f, "return")
-        } else {
-            write!(f, "return {}", self.expression)
-        }
-    }
-}
-
-impl fmt::Display for LoopControlTerm {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            LoopControlTerm::Break => write!(f, "break"),
-            LoopControlTerm::Continue => write!(f, "continue"),
-        }
     }
 }
