+++ compiler/hash-typecheck/src/operations/blocks.rs
@@ -4,8 +4,10 @@ use hash_storage::store::{statics::StoreId, TrivialSequenceStoreKey};
 use hash_tir::{
     context::{HasContext, ScopeKind},
     intrinsics::definitions::never_ty,
-    scopes::{BlockStatement, BlockTerm},
-    tir::{NodeId, NodeOrigin, TermId, Ty, TyId},
+    tir::{
+        blocks::{BlockStatement, BlockTerm},
+        NodeId, NodeOrigin, TermId, Ty, TyId,
+    },
 };
 use hash_utils::log::info;
